requirements.txt
openai
google-serp-api
tiktoken
trafilatura
google-search-results
python-dotenv